pyproject.toml
[project]
name = "telegram-claude-bot"
version = "1.0.0"
description = "Telegram bot interface for Claude Code CLI"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
telegram-claude-bot = "src.main:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
